Avalon bug fixes (#6)
if the default option for a phase is set due to the AI inputting invalid inputs, we set made invalid_move to false.

This is to prevent the phase getting stuck as before, invalid_move was true, and so it would keep redoing the phase, keeping the default output when the AI would get it wrong.

When the default option is set manually after incorrect moves from the AI, the phase advances correctly